Add YOLO mode
This disables the check that the timer was started. Instead of an
exception, you now get garbage data and risk undefined behaviour due to
signed integer overflow. But you avoid some branching if speed is
important to you.

Also discovered a bug in Timer due to an uninitialized variable and a
missing test for throwing exceptions.