Check return value from clock_gettime unless YOLO mode
This "can't fail" since clock_gettime only fails if the clock type is
not supported or the timespec out param is outside the accessible
address space. But let's check it anyway just in case someone tries to
run it on a weird system, has corrupt memory or something like that. But
disable it in YOLO mode.